requirements.txt
Pillow
base64